+++ CardanoSharp.Wallet/CIPs/CIP2/CoinSelectionStrategies/RandomImproveStrategy.cs
@@ -16,13 +16,13 @@ namespace CardanoSharp.Wallet.CIPs.CIP2
     
     public class RandomImproveStrategy: BaseSelectionStrategy, IRandomImproveStrategy
     {
-        public void SelectInputs(CoinSelection coinSelection, List<Utxo> availableUTxOs, ulong amount, Asset asset = null, int limit = 20)
+        public void SelectInputs(CoinSelection coinSelection, List<Utxo> availableUTxOs, long amount, Asset asset = null, int limit = 20)
         {
             //1. Randomly select UTxOs
             var rand = new Random();
             
             //determine
-            ulong currentAmount = GetCurrentBalance(coinSelection, asset);
+            long currentAmount = GetCurrentBalance(coinSelection, asset);
             
             //reorder list
             availableUTxOs = OrderUTxOsByDescending(availableUTxOs, asset);
@@ -53,7 +53,7 @@ namespace CardanoSharp.Wallet.CIPs.CIP2
                 
                 // get quantity of UTxO
                 var quantity = (asset is null)
-                    ? randomUTxO.Balance.Lovelaces
+                    ? (long)randomUTxO.Balance.Lovelaces
                     : randomUTxO.Balance.Assets.FirstOrDefault(x => x.PolicyId.SequenceEqual(asset.PolicyId) && x.Name.Equals(asset.Name)).Quantity;
 
                 // increment current amount by the UTxO quantity
@@ -86,14 +86,14 @@ namespace CardanoSharp.Wallet.CIPs.CIP2
         }
 
         //why is this static
-        private static bool CalculateCondition(Utxo v1, ulong ideal, ulong max, Utxo v0, int limit, ICollection utxos, Asset asset)
+        private static bool CalculateCondition(Utxo v1, long ideal, long max, Utxo v0, int limit, ICollection utxos, Asset asset)
         {
             bool[] arrayToMatchConditions = { false, false, false,};
-            arrayToMatchConditions[0] =  asset is null ? Math.Abs((long) (ideal - v1.Balance.Lovelaces)) < Math.Abs((long) (ideal -  (v0.Balance?.Lovelaces ?? 0))) 
+            arrayToMatchConditions[0] =  asset is null ? Math.Abs((long) (ideal - (long)v1.Balance.Lovelaces)) < Math.Abs((long) (ideal -  (long)(v0.Balance?.Lovelaces ?? 0))) 
                 : Math.Abs((long) (ideal - v1.Balance.Assets.FirstOrDefault(x => x.PolicyId.SequenceEqual(asset.PolicyId) && x.Name.Equals(asset.Name)).Quantity)) < 
                   Math.Abs((long) (ideal - (v0.Balance?.Assets?.FirstOrDefault(x => x.PolicyId.SequenceEqual(asset.PolicyId) && x.Name.Equals(asset.Name))?.Quantity ?? 0))) ;
 
-            arrayToMatchConditions[1] = asset is null ? v1.Balance.Lovelaces <= max : v1.Balance.Assets.FirstOrDefault(x => x.PolicyId.SequenceEqual(asset.PolicyId) && x.Name.Equals(asset.Name)).Quantity <= max;
+            arrayToMatchConditions[1] = asset is null ? (long)v1.Balance.Lovelaces <= max : v1.Balance.Assets.FirstOrDefault(x => x.PolicyId.SequenceEqual(asset.PolicyId) && x.Name.Equals(asset.Name)).Quantity <= max;
 
             //TODO how do we define the maximum input count as listed below
             //Condition 3: when counting cumulatively across all outputs considered so far, we have not selected more than the maximum number of UTxO entries specified by Maximum Input Count.
